refactor: 更新仓库包名
