+++ ios/WarehouseManagementSystem/Images.xcassets/AppIcon.appiconset/Contents.json
