+++ ios/WarehouseManagementSystem/LaunchScreen.storyboard
