+++ ios/WarehouseManagementSystem/PrivacyInfo.xcprivacy
