+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ProcureSellManage",
+  "name": "warehouse-management-system",
   "version": "2.0.1",
   "versionCode": 3,
   "buildNumber": "201",
